backend/requirements_dev.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.10.0
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
deprecation==2.1.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.116.1
Flask==3.1.2
fonttools==4.55.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.14.1
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
panda==0.3.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
postgrest==1.1.1
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pyjokes==0.8.3
PyJWT==2.10.1
pyparsing==3.2.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyttsx3==2.97
pytz==2024.2
pywin32==306
pyzmq==26.2.0
realtime==2.7.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9.1
scipy==1.15.2
setuptools==75.5.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
storage3==0.12.1
streamlit==1.40.1
StrEnum==0.4.15
supabase==2.18.0
supafunc==0.10.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
